Add token launch and apply sections to new profile (#727)
