+++ app/packs/src/components/profile/Token.jsx
@@ -8,13 +8,17 @@ import Button from "src/components/design_system/button";
 import { useWindowDimensionsHook } from "src/utils/window";
 import { useTheme } from "src/contexts/ThemeContext";
 
-import { formatNumberWithSymbol } from "src/utils/viewHelpers";
+import { formatNumberWithSymbol, shortenAddress } from "src/utils/viewHelpers";
 
 import cx from "classnames";
 
-const Token = ({ tokenData, talent, env }) => {
+const Token = ({ talent, talentTokenPrice, railsContext }) => {
   const { mobile } = useWindowDimensionsHook();
   const { mode } = useTheme();
+  const token = talent.token;
+
+  const totalSupply = ethers.utils.formatUnits(talent.totalSupply);
+  const maxSupply = ethers.utils.formatUnits(talent.maxSupply);
 
   const addTokenToMetamask = async () => {
     const onChainAPI = new OnChain(railsContext.contractsEnv);
@@ -22,13 +26,10 @@ const Token = ({ tokenData, talent, env }) => {
     const chainId = await onChainAPI.getChainID();
 
     if (chainId != token.chainId) {
-      await onChainAPI.switchChain(talent.token.chainId);
+      await onChainAPI.switchChain(token.chainId);
     }
 
-    await onChainAPI.addTokenToWallet(
-      talent.token.contract_id,
-      talent.token.ticker
-    );
+    await onChainAPI.addTokenToWallet(token.contractId, token.ticker);
   };
 
   return (
@@ -41,18 +42,16 @@ const Token = ({ tokenData, talent, env }) => {
       <div className="row">
         <div className={cx("col-12 col-lg-4", mobile && "mb-6")}>
           <H3
-            text={formatNumberWithSymbol(
-              tokenData.totalSupply * tokenData.price
-            )}
+            text={formatNumberWithSymbol(totalSupply)}
             className="text-center inverted-text-primary-01"
           ></H3>
           <P3 className="text-center inverted-text-primary-03">Market Cap</P3>
         </div>
         <div className={cx("col-12 col-lg-4", mobile && "mb-6")}>
           <H3
-            text={`${ethers.utils.commify(
-              Number.parseFloat(tokenData.totalSupply).toFixed(2)
-            )} $${talent.token.ticker}`}
+            text={`${ethers.utils.commify(totalSupply * talentTokenPrice)} $${
+              token.ticker
+            }`}
             className="text-center inverted-text-primary-01"
           ></H3>
           <P3 className="text-center inverted-text-primary-03">
@@ -61,9 +60,7 @@ const Token = ({ tokenData, talent, env }) => {
         </div>
         <div className={cx("col-12 col-lg-4", mobile && "mb-6")}>
           <H3
-            text={formatNumberWithSymbol(
-              ethers.utils.formatUnits(talent.maxSupply)
-            )}
+            text={formatNumberWithSymbol(maxSupply)}
             className="text-center inverted-text-primary-01"
           ></H3>
           <P3 className="text-center inverted-text-primary-03">Max Supply</P3>
@@ -72,7 +69,7 @@ const Token = ({ tokenData, talent, env }) => {
       <div className={cx("row", mobile ? "" : "mt-7")}>
         <div className={cx("col-12 col-lg-4", mobile && "mb-6")}>
           <H3
-            text={talent.supportersCount}
+            text={talent.supportersCount || "0"}
             className="text-center inverted-text-primary-01"
           ></H3>
           <P3 className="text-center inverted-text-primary-03">Supporters</P3>
@@ -88,25 +85,26 @@ const Token = ({ tokenData, talent, env }) => {
         </div>
         <div className={cx("col-12 col-lg-4", mobile && "mb-6")}>
           <H3
-            text={formatNumberWithSymbol(
-              tokenData.totalSupply * tokenData.price
-            )}
+            text={shortenAddress(token.contractId)}
             className="text-center inverted-text-primary-01"
           ></H3>
           <P3
-            text={`${chainIdToName(talent.token.chainId, env)} Network`}
+            text={`${chainIdToName(
+              token.chainId,
+              railsContext.contractsEnv
+            )} Network`}
             className="text-center inverted-text-primary-03"
           ></P3>
         </div>
       </div>
       <div className="d-flex flex-column justify-content-center mt-6">
         <Button
-          className="mr-2 mt-2 mx-auto add-to-metamask"
+          className="mr-2 mt-2 mx-auto inverted-button"
           mode={mode() == "light" ? "dark" : "light"}
           type="white-default"
           onClick={() => addTokenToMetamask()}
         >
-          <P2 bold text={`Add $${talent.token.ticker} to Metamask`} />
+          <P2 bold text={`Add $${token.ticker} to Metamask`} />
         </Button>
       </div>
     </section>
