+++ app/packs/src/components/profile/show.jsx
@@ -7,13 +7,6 @@ import { camelCaseObject } from "src/utils/transformObjects";
 import { useWindowDimensionsHook } from "src/utils/window";
 import Button from "src/components/design_system/button";
 
-import {
-  ApolloProvider,
-  useQuery,
-  GET_TALENT_PORTFOLIO_FOR_ID_SIMPLE,
-  client,
-} from "src/utils/thegraph";
-
 import Poaps from "./web3/poaps";
 import Nfts from "./web3/nfts";
 import Tokens from "./web3/tokens";
@@ -22,22 +15,17 @@ import About from "./About";
 import Journey from "./Journey";
 import Community from "./Community";
 import Token from "./Token";
+import LaunchToken from "./LaunchToken";
+import ApplyToLaunchToken from "./ApplyToLaunchToken";
 
 const Show = ({ talent, railsContext, currentUserId }) => {
   const [localTalent, setLocalTalent] = useState(camelCaseObject(talent));
+  const user = localTalent.user;
+  const token = localTalent.token;
   const canUpdate = talent.user.id == currentUserId;
   const [selectedSection, setSelectedSection] = useState(window.location.hash);
-  const [listLoaded, setListLoaded] = useState(false);
   const [showLastDivider, setShowLastDivider] = useState(false);
-  const [tokenData, setTokenData] = useState({
-    price: 0.1,
-    totalSupply: 0,
-  });
-
-  const { loading, data } = useQuery(GET_TALENT_PORTFOLIO_FOR_ID_SIMPLE, {
-    variables: { id: localTalent.token.contractId?.toLowerCase() },
-    skip: listLoaded,
-  });
+  const talentTokenPrice = 0.1;
 
   const { mobile } = useWindowDimensionsHook();
 
@@ -67,20 +55,22 @@ const Show = ({ talent, railsContext, currentUserId }) => {
     }
   }, [selectedSection]);
 
-  useEffect(() => {
-    if (loading || !data || !data.talentToken) {
-      if (!loading) {
-        setListLoaded(true);
-      }
-      return;
+  const getCurrentTokenSection = () => {
+    if (
+      (user.profileType == "approved" || user.profileType == "talent") &&
+      !token.contractId
+    ) {
+      return LaunchToken;
     }
 
-    setListLoaded(true);
-    setTokenData({
-      ...tokenData,
-      totalSupply: ethers.utils.formatUnits(data.talentToken.totalSupply || 0),
-    });
-  }, [data, loading]);
+    if (user.profileType != "approved" && user.profileType != "talent") {
+      return ApplyToLaunchToken;
+    }
+
+    return Token;
+  };
+
+  const CurrentTokenSection = getCurrentTokenSection();
 
   return (
     <div className="d-flex flex-column lg-h-100 p-0 mt-7">
@@ -88,10 +78,10 @@ const Show = ({ talent, railsContext, currentUserId }) => {
         className="mb-2"
         talent={localTalent}
         setTalent={setLocalTalent}
-        tokenData={tokenData}
         currentUserId={currentUserId}
         railsContext={railsContext}
         changeSection={changeSection}
+        talentTokenPrice={talentTokenPrice}
       />
       <Divider className="my-6" />
       <div className="d-flex justify-content-lg-center overflow-x-scroll mx-4">
@@ -109,9 +99,9 @@ const Show = ({ talent, railsContext, currentUserId }) => {
         />
         <Button
           className="mr-2"
-          type={buttonType(`#${localTalent.token.ticker}`)}
-          text={localTalent.token.ticker}
-          onClick={() => changeSection(`#${localTalent.token.ticker}`)}
+          type={buttonType(`#${token.ticker}`)}
+          text={token.ticker}
+          onClick={() => changeSection(`#${token.ticker}`)}
         />
         <Button
           className="mr-2"
@@ -132,11 +122,13 @@ const Show = ({ talent, railsContext, currentUserId }) => {
       <div className="my-7 w-100" id="#Journey">
         <Journey talent={localTalent} />
       </div>
-      <div className="my-7 w-100" id={`#${localTalent.token.ticker}`}>
-        <Token
-          tokenData={tokenData}
+      <div className="my-7 w-100" id={"#token"}>
+        <CurrentTokenSection
           talent={localTalent}
-          env={railsContext.contractsEnv}
+          talentTokenPrice={talentTokenPrice}
+          setLocalTalent={setLocalTalent}
+          railsContext={railsContext}
+          waitingApproval={user.profileType == "waiting_for_approval"}
         />
       </div>
       <div className="my-7 w-100" id="#Community">
@@ -171,9 +163,7 @@ const Show = ({ talent, railsContext, currentUserId }) => {
 export default (props, railsContext) => {
   return () => (
     <ThemeContainer>
-      <ApolloProvider client={client(props.talent.token.chain_id)}>
-        <Show {...props} railsContext={railsContext} />
-      </ApolloProvider>
+      <Show {...props} railsContext={railsContext} />
     </ThemeContainer>
   );
 };
